@@ -1,10 +1,11 @@
 buildscript {
-    ext.kotlin_version = '0.11.91.1'
+    ext.kotlin_version = '1.9.25'
     repositories {
+        google()
         mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:1.0.0'
+        classpath 'com.android.tools.build:gradle:8.2.2'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
@@ -13,20 +14,21 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 
 dependencies {
-    compile project(':androidpagecontrol')
-    compile 'com.android.support:support-v4:20.0.0'
-    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    implementation project(':androidpagecontrol')
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 }
 
 android {
-    compileSdkVersion 20
-    buildToolsVersion "20.0.0"
+    namespace 'com.androidpagecontrol.samples.demos'
+    compileSdk 34
 
     defaultConfig {
+        applicationId "com.androidpagecontrol.samples.demos"
         versionCode 1
         versionName VERSION_NAME
-        minSdkVersion 8
-        targetSdkVersion 20
+        minSdk 16
+        targetSdk 34
     }
 
     sourceSets {
