@@ -1,10 +1,11 @@
 buildscript {
-    ext.kotlin_version = '0.11.91.1'
+    ext.kotlin_version = '1.9.25'
     repositories {
+        google()
         mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:1.0.0'
+        classpath 'com.android.tools.build:gradle:8.2.2'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
@@ -13,17 +14,18 @@ apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
 
 dependencies {
-    compile 'com.android.support:support-v4:20.0.0'
-    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.viewpager:viewpager:1.0.0'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 }
 
 android {
-    compileSdkVersion 20
-    buildToolsVersion "20.0.0"
+    namespace 'com.androidpagecontrol'
+    compileSdk 34
 
     defaultConfig {
-        minSdkVersion 8
-        targetSdkVersion 20
+        minSdk 16
+        targetSdk 34
     }
 
     sourceSets {
@@ -33,6 +35,28 @@ android {
     lintOptions {
         abortOnError false
     }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
+    }
+
+    kotlinOptions {
+        jvmTarget = '17'
+    }
+
+    publishing {
+        singleVariant("release") {
+            withSourcesJar()
+            withJavadocJar()
+        }
+    }
+}
+
+ext {
+    POM_NAME = 'Android-PageControl'
+    POM_ARTIFACT_ID = 'androidpagecontrol'
+    POM_PACKAGING = 'aar'
 }
 
 apply from: '../kotlin-publish.gradle'
